refactor: convert server from hook to common function
